+++ app/src/main/java/org/Traffix/animations/FonctionFinAnimation.java
@@ -1,4 +1,4 @@
-package animations;
+package org.Traffix.animations;
 
 public abstract class FonctionFinAnimation{
     public abstract void appeler();
