+++ app/src/main/java/org/Traffix/animations/GestionnaireAnimations.java
@@ -1,4 +1,4 @@
-package animations;
+package org.Traffix.animations;
 
 import java.util.ArrayList;
