+++ app/src/main/java/org/Traffix/maths/Vec3.java
@@ -1,118 +1,115 @@
-package org.example.maths;
+package org.Traffix.maths;
 
 import animations.Animable;
 
 /**
  * Vecteur de 3 dimension à coordonées cartésiennes (x,y,z)
  */
-public class Vec4 implements Animable{
+public class Vec3 implements Animable {
     public float x;
     public float y;
     public float z;
-    public float w;
 
     /**
      * Créé un nouveau vecteur
      * @param x - La composante x
      * @param y - La composante y
      * @param z - La composante z
-     * @param w - La composante w
      */
-    public Vec4(float x, float y, float z, float w){
+    public Vec3(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
-        this.w = w;
     }
 
     /**
      * Créé un nouveau vecteur en clonant le vecteur clone
      * @param clone
      */
-    public Vec4(Vec4 clone){
+    public Vec3(Vec3 clone){
         this.x = clone.x;
         this.y = clone.y;
         this.z = clone.z;
-        this.w = clone.w;
     }
 
     /**
      * Créé un nouveau vecteur
      * @param xyz - Composantes x,y,z
      */
-    public Vec4(float xyzw){
-        this.x = xyzw;
-        this.y = xyzw;
-        this.z = xyzw;
-        this.w = xyzw;
+    public Vec3(float xyz){
+        this.x = xyz;
+        this.y = xyz;
+        this.z = xyz;
     }
 
     /**
      * Additionne b à ce vecteur (a+b)
      * @param b - Vecteur à additionner
      */
-    public void addi(Vec4 b){
+    public Vec3 addi(Vec3 b){
         x += b.x;
         y += b.y;
         z += b.z;
-        w += b.w;
+        return this;
     }
 
     /**
      * Soustrait b à ce vecteur (a-b)
      * @param b - Vecteur à soustraire
      */
-    public void sous(Vec4 b){
+    public Vec3 sous(Vec3 b){
         x -= b.x;
         y -= b.y;
         z -= b.z;
-        w -= b.w;
+        return this;
     }
 
     /**
      * Multiplie ce vecteur par le scalaire s (a*s)
      * @param s - Facteur multiplicateur
      */
-    public void mult(float s){
+    public Vec3 mult(float s){
         x *= s;
         y *= s;
         z *= s;
-        w *= s;
+        return this;   
     }
 
     /**
      * Multiplie les composantes de ce vecteur par les composantes de m (a.x*m.x; a.y*m.y)
      * @param m - Vecteur multiplicateur
      */
-    public void mult(Vec4 m){
+    public Vec3 mult(Vec3 m){
         x *= m.x;
         y *= m.y;
         z *= m.z;
+        return this;
     }
 
     /**
      * Divise les composantes de ce vecteur par les composantes de d (a.x/d.x; a.y/d.y)
      * @param d
      */
-    public void div(Vec4 d){
+    public Vec3 div(Vec3 d){
         x = x/d.x;
         y = y/d.y;
         z = z/d.z;
-        w = w/d.w;
+        return this;
     }
 
     /**
      * Normalise ce vecteur
      */
-    public void norm(){
-        if(longueur() > 0){
-            x = x/ longueur();
-            y = y/ longueur();
-            z = z/ longueur();
-            w = w/ longueur();
+    public Vec3 norm(){
+        float l = longueur();
+        if(l > 0){
+            x = x/ l;
+            y = y/ l;
+            z = z/ l;
         }else{
             System.err.println("Normalisation de vecteur nul. Les composantes resteront 0");
         }
+        return this;
     }
 
 
@@ -121,7 +118,20 @@ public class Vec4 implements Animable{
      * @return ||a||
      */
     public float longueur(){
-        return (float) Math.sqrt(x*x+y*y+z*z+w*w);
+        return (float) Math.sqrt(x*x+y*y+z*z);
+    }
+
+    /**
+     * Retourne le produit vectoriel entre ce vecteur et c (a^c)
+     * @param c
+     * @return a^c
+     */
+    public Vec3 croix(Vec3 c){
+        Vec3 r = new Vec3(0,0,0);
+        r.x = (y*c.z)-(z*c.y);
+        r.y = (z*c.x)-(x*c.z);
+        r.z = (x*c.y)-(y*c.x);
+        return r;
     }
 
     /**
@@ -130,8 +140,8 @@ public class Vec4 implements Animable{
      * @param b
      * @return ||a-b||
      */
-    public static float distance(Vec4 a, Vec4 b){
-        return (float) Math.sqrt((a.x-b.x) * (a.x-b.x) + (a.y-b.y) * (a.y-b.y) + (a.z-b.z) * (a.z-b.z) + (a.w-b.w) * (a.w-b.w));
+    public static float distance(Vec3 a, Vec3 b){
+        return (float) Math.sqrt((a.x-b.x) * (a.x-b.x) + (a.y-b.y) * (a.y-b.y) + (a.z-b.z) * (a.z-b.z));
     }
 
     /**
@@ -140,8 +150,8 @@ public class Vec4 implements Animable{
      * @param b
      * @return a+b
      */
-    public static Vec4 addi(Vec4 a, Vec4 b){
-        return new Vec4(a.x+b.x, a.y + b.y, a.z+b.z, a.w+b.w);
+    public static Vec3 addi(Vec3 a, Vec3 b){
+        return new Vec3(a.x+b.x, a.y + b.y, a.z+b.z);
     }
 
     /**
@@ -150,8 +160,8 @@ public class Vec4 implements Animable{
      * @param b
      * @return a-b
      */
-    public static Vec4 sous(Vec4 a, Vec4 b){
-        return new Vec4(a.x-b.x, a.y - b.y, a.z-b.z, a.w-b.w);
+    public static Vec3 sous(Vec3 a, Vec3 b){
+        return new Vec3(a.x-b.x, a.y - b.y, a.z-b.z);
     }
 
     /**
@@ -160,8 +170,8 @@ public class Vec4 implements Animable{
      * @param s
      * @return a*s
      */
-    public static Vec4 mult(Vec4 a, float s){
-        return new Vec4(a.x*s, a.y*s, a.z*s, a.w*s);
+    public static Vec3 mult(Vec3 a, float s){
+        return new Vec3(a.x*s, a.y*s, a.z*s);
     }
 
     /**
@@ -170,8 +180,8 @@ public class Vec4 implements Animable{
      * @param b
      * @return (a.x*b.x; a.y*b.y)
      */
-    public static Vec4 mult(Vec4 a, Vec4 b){
-        return new Vec4(a.x*b.x,a.y*b.y, a.z*b.z, a.w*b.w);
+    public static Vec3 mult(Vec3 a, Vec3 b){
+        return new Vec3(a.x*b.x,a.y*b.y, a.z*b.z);
     }
 
     /**
@@ -180,8 +190,8 @@ public class Vec4 implements Animable{
      * @param b
      * @return (a.x/b.x; a.y/b.y)
      */
-    public static Vec4 div(Vec4 a, Vec4 b){
-        return new Vec4(a.x/b.x,a.y/b.y, a.z/b.z, a.w/b.w);
+    public static Vec3 div(Vec3 a, Vec3 b){
+        return new Vec3(a.x/b.x,a.y/b.y, a.z/b.z);
     }
 
     /**
@@ -190,8 +200,36 @@ public class Vec4 implements Animable{
      * @param b
      * @return a•
      */
-    public static float scal(Vec4 a, Vec4 b){
-        return (a.x*b.x)+(a.y*b.y)+(a.z*b.z)+(a.w*b.w);
+    public static float scal(Vec3 a, Vec3 b){
+        return (a.x*b.x)+(a.y*b.y)+(a.z*b.z);
+    }
+
+    /**
+     * Retourne le produit vectoriel entre les vecteurs a et b
+     * @param a
+     * @param b
+     * @return a^b
+     */
+    /* |i  j  k |
+     * |ax ay az|    |ay az|    |ax az|    |ax ay|
+     * |bx by bz| = i|by bz| - j|bx bz| + k|bx by| = i(ay*bz-az*by) - j(ax*bz-az*bx) + k(ax*by-ay*bx)
+     * */
+    public static Vec3 croix(Vec3 a, Vec3 b){
+        return new Vec3(a.y*b.z - a.z*b.y, -(a.x*b.z - a.z*b.x), a.x*b.y - a.y*b.x);
+    }
+    /**
+     * Renvoi le produit mixte entre les vecteur a, b et c
+     * @param a  
+     * @param b
+     * @param c
+     * @return a•(b^c)
+     */
+    /* |ax ay az|
+     * |bx by bz|     |by bz|     |bx bz|     |bx by|
+     * |cx cy cz| = ax|cy cz| + ay|cx cz| + az|cx cy| = ax(by*cz-bz*cy) + ay(bx*cz-bz*cx) + az(bx*cy-by*cx)
+     * */
+    public static float mixte(Vec3 a, Vec3 b, Vec3 c){
+        return  a.x*(b.y*c.z-b.z*c.y) + a.y*(b.z*c.x-b.x*c.z) + a.z*(b.x*c.y-b.y*c.x);
     }
 
     /**
@@ -199,10 +237,9 @@ public class Vec4 implements Animable{
      * @param a
      * @return (1/||a||)*a
      */
-    public static Vec4 norm(Vec4 a){
+    public static Vec3 norm(Vec3 a){
         if(a.longueur() > 0){
-        	float l = a.longueur();
-            return new Vec4(a.x/l,a.y/l,a.z/l,a.z/l);
+            return new Vec3(a.x/a.longueur(),a.y/a.longueur(),a.z/a.longueur());
         }else{
             //System.err.println("Normalisation d'un vecteur nul. Les composantes resteront 0.");
             return a;
@@ -216,10 +253,10 @@ public class Vec4 implements Animable{
      * @param assiette
      * @return (angleX, angleY, angleZ) qui pointe vers 'positionInitiale' à partir de 'destination' avec une assiette de 'assiette'
      */
-    public static Vec4 dirigerVers(Vec4 positionInitiale, Vec4 destination, float assiette){
-        Vec4 a = Vec4.sous(positionInitiale,destination);
-        Vec4 b = new Vec4(0,0,0,0);
-        Vec4 c = Vec4.norm(a);
+    public static Vec3 dirigerVers(Vec3 positionInitiale, Vec3 destination, float assiette){
+        Vec3 a = Vec3.sous(positionInitiale,destination);
+        Vec3 b = new Vec3(0,0,0);
+        Vec3 c = Vec3.norm(a);
         b.x = (float)Math.toDegrees(Math.asin(-c.y));
         b.y = (float)Math.toDegrees(Math.atan2(a.x,a.z));
         b.z = assiette;
@@ -230,34 +267,41 @@ public class Vec4 implements Animable{
      * Retourne une copie de ce vecteur
      * @return copie de a
      */
-    public Vec4 copier(){
-        return new Vec4(x,y,z,w);
+    public Vec3 copier(){
+        return new Vec3(x,y,z);
     }
     
     /**
      * Retourne le vecteur opposé à ce vecteur
      * @return -a
      */
-    public Vec4 neg(){
-        return new Vec4(-x,-y,-z,-w);
+    public Vec3 opposé(){
+        return new Vec3(-x,-y,-z);
     }
 
+    /**
+     * Retourne le vecteur inverse de ce vecteur
+     * @return 1/a
+     */
+    public Vec3 inv(){
+        return new Vec3(1f/x,1f/y,1f/z);
+    }
+    
     @Override
     public void mix(Object[] a, Object[] b, float m) {
         x = (float)a[0]*(1f-m) + (float)b[0]*m;
         y = (float)a[1]*(1f-m) + (float)b[2]*m;
         z = (float)a[2]*(1f-m) + (float)b[2]*m;
-        w = (float)a[3]*(1f-m) + (float)b[3]*m;
     }
 
     @Override
     public Object[] animClé() {
-        return new Object[]{x,y,z,w};
+        return new Object[]{x,y,z};
     }
 
     @Override
     public boolean validerClé(Object[] c) {
-        if (c.length != 4 || !(c[0] instanceof Float) || !(c[1] instanceof Float) || !(c[2] instanceof Float) || !(c[3] instanceof Float)){
+        if (c.length != 3 || !(c[0] instanceof Float) || !(c[1] instanceof Float) || !(c[2] instanceof Float)){
             return false;
         }
         return true;
@@ -268,6 +312,6 @@ public class Vec4 implements Animable{
         x = (float)cléB[0];
         y = (float)cléB[1];
         z = (float)cléB[2];
-        w = (float)cléB[3];
     }
+
 }
