+++ app/src/test/java/org/example/AppTest.java
@@ -6,6 +6,8 @@ package org.example;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
+import org.Traffix.App;
+
 public class AppTest {
     @Test public void appHasAGreeting() {
         App classUnderTest = new App();
